requirements.txt
pytz
requests
python-dotenv
ipython
bs4

# Django
django
django-timezone-field 
django-cors-headers
django-debug-toolbar
whitenoise
Pillow

# Django 3.0 async requirements
gunicorn
uvicorn
uvloop
httptools

# Databases
psycopg2-binary
dj-database-url

# Distributed tasks
celery
redis
django-celery-beat

# Telegram
python-telegram-bot
#telebot

# monitoring
# sentry-sdk
